+++ src/sim/VLayersProcessor.h
@@ -65,8 +65,8 @@ const VSimTriangle::const_vector_ptr &getActiveTriangles() const ;
 bool isLayerVisible(unsigned int layer) const ;
 bool isLayerEnabled(unsigned int layer) const ;
 
-void setInjectionPoint(const QVector3D &point, float diameter);
-void setVacuumPoint(const QVector3D &point, float diameter);
+void setInjectionPoint(const QVector3D &point, double diameter);
+void setVacuumPoint(const QVector3D &point, double diameter);
 
 private:
     void updateActiveElementsVectors() ;
